refactoring: use virtual methods and inheritance in php-queries.cpp
